# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlipStudio Animation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/duplicate:
    post:
      operationId: duplicateProject
      tags: [projects]
      summary: Duplicate a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "201":
          description: Duplicated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}/thumbnail:
    put:
      operationId: updateProjectThumbnail
      tags: [projects]
      summary: Update project thumbnail (base64 PNG)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThumbnailUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}/layers:
    get:
      operationId: listLayers
      tags: [layers]
      summary: List layers for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Layers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Layer"
    post:
      operationId: createLayer
      tags: [layers]
      summary: Add a layer
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerInput"
      responses:
        "201":
          description: Created layer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Layer"
  /projects/{projectId}/layers/{layerId}:
    patch:
      operationId: updateLayer
      tags: [layers]
      summary: Update a layer
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/LayerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerUpdate"
      responses:
        "200":
          description: Updated layer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Layer"
    delete:
      operationId: deleteLayer
      tags: [layers]
      summary: Delete a layer
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/LayerId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/layers/reorder:
    patch:
      operationId: reorderLayers
      tags: [layers]
      summary: Reorder layers
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayerReorder"
      responses:
        "200":
          description: Reordered layers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Layer"
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: integer }
...
    LayerId:
      in: path
      name: layerId
      required: true
      schema: { type: integer }
  schemas:
...
    Project:
      type: object
      required: [id, name, fps, canvasWidth, canvasHeight, frameCount, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        fps: { type: integer }
        canvasWidth: { type: integer }
        canvasHeight: { type: integer }
        frameCount: { type: integer }
        thumbnailData: { type: ["string", "null"] }
        backgroundColor: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectDetail:
      type: object
      required: [id, name, fps, canvasWidth, canvasHeight, frameCount, frames, layers, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        fps: { type: integer }
        canvasWidth: { type: integer }
        canvasHeight: { type: integer }
        frameCount: { type: integer }
        thumbnailData: { type: ["string", "null"] }
        backgroundColor: { type: string }
        frames:
          type: array
          items:
            $ref: "#/components/schemas/Frame"
        layers:
          type: array
          items:
            $ref: "#/components/schemas/Layer"
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        fps: { type: integer, default: 12 }
        canvasWidth: { type: integer, default: 1920 }
        canvasHeight: { type: integer, default: 1080 }
        backgroundColor: { type: string, default: "#ffffff" }
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        fps: { type: integer }
        backgroundColor: { type: string }
    ThumbnailUpdate:
      type: object
      required: [thumbnailData]
      properties:
        thumbnailData: { type: string }
    Frame:
      type: object
      required: [id, projectId, frameIndex, duration, isHold, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        frameIndex: { type: integer }
        duration: { type: integer }
        isHold: { type: boolean }
        canvasData: { type: ["string", "null"] }
        thumbnailData: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Layer:
      type: object
      required: [id, projectId, name, layerIndex, isVisible, isLocked, opacity, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        frameId: { type: ["integer", "null"] }
        name: { type: string }
        layerIndex: { type: integer }
        isVisible: { type: boolean }
        isLocked: { type: boolean }
        opacity: { type: number }
        blendMode: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    LayerInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        frameId: { type: integer }
        layerIndex: { type: integer }
        opacity: { type: number, default: 1.0 }
        blendMode: { type: string, default: "normal" }
    LayerUpdate:
      type: object
      properties:
        name: { type: string }
        isVisible: { type: boolean }
        isLocked: { type: boolean }
        opacity: { type: number }
        blendMode: { type: string }
    LayerReorder:
      type: object
      required: [layerIds]
      properties:
        layerIds:
          type: array
          items: { type: integer }
...