# lib/api-zod/src/generated/types/layer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Layer {
  id: number;
  projectId: number;
  /** @nullable */
  frameId?: number | null;
  name: string;
  layerIndex: number;
  isVisible: boolean;
  isLocked: boolean;
  opacity: number;
  blendMode?: string;
  createdAt: string;
  updatedAt: string;
}