# lib/api-zod/src/generated/types/layerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface LayerInput {
  name: string;
  frameId?: number;
  layerIndex?: number;
  opacity?: number;
  blendMode?: string;
}