# lib/api-zod/src/generated/types/layerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface LayerUpdate {
  name?: string;
  isVisible?: boolean;
  isLocked?: boolean;
  opacity?: number;
  blendMode?: string;
}