# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { Frame } from "./frame";
import type { Layer } from "./layer";

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  fps: number;
  canvasWidth: number;
  canvasHeight: number;
  frameCount: number;
  /** @nullable */
  thumbnailData?: string | null;
  backgroundColor?: string;
  frames: Frame[];
  layers: Layer[];
  createdAt: string;
  updatedAt: string;
}